requirements.txt
Mastodon.py
python-dotenv
##
zhtts
##
# git+https://github.com/fchollet/stable-diffusion-tensorflow
# git+https://github.com/robrohan/stable-diffusion-tensorflow
/home/rob/Projects/home-lab/stable-diffusion-tensorflow
tensorflow==2.9.0
h5py==3.7.0
Pillow==9.2.0
tqdm==4.64.1
ftfy==6.1.1
regex==2022.9.13
tensorflow-addons==0.17.1